move menu to it's own component
